# Basic SAX interface with benchmarks