… benchmark/CMakeLists.txt
… benchmark/bench_dom_api.cpp
… benchmark/bench_sax.cpp
… benchmark/twitter/sax_tweet_reader.h
… benchmark/twitter/sax_tweet_reader_visitor.h
… benchmark/twitter/tweet.h
… benchmark/twitter/twitter_user.h
… src/generic/stage2/numberparsing.h
… src/generic/stage2/stringparsing.h
… src/generic/stage2/structural_parser.h
… src/generic/stage2/tape_builder.h
… src/haswell/dom_parser_implementation.cpp
… src/implementation.cpp
… src/westmere/dom_parser_implementation.cpp