This is as far as I could push the SAX interface without affecting performance. Pretty much moving *any* line seems to be enough to hurt at least some files by 1-5% overall. 

It makes no functional changes. What it does:
* Removes structural_parser and structural_iterator, replacing them with a single, fairly thin json_iterator class. walk_document() was pretty much the only method left in structural_parser.
* Give visitor (tape_builder) methods the opportunity to return error codes
* Move more functionality into the visitor
* Make walk_document a little easier to read (IMO, we'll see what you think :))

What I *really* like, as a non-speculative benefit for our current code in particular, is the clean separation of concerns. `json_iterator` owns everything necessary to iterate (structural iteration, depth and is_array), and `tape_builder` owns everything related to parsing to tape (tape index, container indexes, and the string buffer). Hopefully it helps us when making changes.

There is a little bit of API documentation now, but it's not at full speed.

That makes it a refactor-only change though, which means performance must not suffer!

## Performance

skylake sees mixed impact, with slightly increased instruction count generally:

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |
| numbers                          |   2345 | 266.9 | 251.7 |    5% | 819.3 | 836.4 |   -2% |
| mesh                             |  11306 | 308.9 | 297.4 |    3% | 1009.0 | 1016.0 |    0% |
| marine_ik                        |  46616 | 294.5 | 285.9 |    2% | 919.6 | 921.4 |    0% |
| github_events                    |   1017 |  79.5 |  78.5 |    1% | 264.5 | 264.5 |    0% |
| random                           |   7976 | 141.6 | 139.9 |    1% | 494.6 | 494.5 |    0% |
| instruments                      |   3442 | 105.2 | 104.3 |    0% | 378.5 | 378.6 |    0% |
| update-center                    |   8330 | 112.0 | 111.4 |    0% | 335.5 | 335.5 |    0% |
| apache_builds                    |   1988 |  85.9 |  85.6 |    0% | 304.7 | 304.7 |    0% |
| mesh.pretty                      |  24646 | 179.5 | 178.9 |    0% | 592.0 | 596.0 |    0% |
| twitterescaped                   |   8787 | 184.5 | 185.9 |    0% | 489.9 | 489.9 |    0% |
| twitter                          |   9867 |  91.7 |  92.5 |    0% | 290.2 | 290.2 |    0% |
| citm_catalog                     |  26987 |  80.2 |  81.5 |   -1% | 296.3 | 296.3 |    0% |
| canada                           |  35172 | 288.8 | 294.1 |   -1% | 973.9 | 989.3 |   -1% |
| gsoc-2018                        |  51997 |  71.8 |  73.6 |   -2% | 167.0 | 167.0 |    0% |

ARM (ampere gcc10.2) is essentially unimpacted, with minute instruction count fluctuations:

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |
| marine_ik                        |  46616 | 861.4 | 859.1 |    0% | 968.3 | 966.6 |    0% |
| github_events                    |   1017 | 461.0 | 459.7 |    0% | 357.0 | 357.9 |    0% |
| numbers                          |   2345 | 888.3 | 886.9 |    0% | 873.9 | 869.7 |    0% |
| canada                           |  35172 | 811.2 | 810.3 |    0% | 996.4 | 993.2 |    0% |
| mesh                             |  11306 | 922.5 | 921.6 |    0% | 1047.7 | 1044.8 |    0% |
| apache_builds                    |   1988 | 474.3 | 473.9 |    0% | 411.5 | 412.4 |    0% |
| update-center                    |   8330 | 553.4 | 553.2 |    0% | 455.2 | 456.7 |    0% |
| gsoc-2018                        |  51997 | 359.6 | 359.9 |    0% | 237.6 | 237.9 |    0% |
| instruments                      |   3442 | 489.5 | 490.0 |    0% | 464.5 | 466.1 |    0% |
| twitterescaped                   |   8787 | 739.4 | 740.8 |    0% | 620.1 | 621.5 |    0% |
| twitter                          |   9867 | 491.0 | 491.9 |    0% | 381.9 | 383.1 |    0% |
| mesh.pretty                      |  24646 | 563.9 | 565.3 |    0% | 624.7 | 623.4 |    0% |
| citm_catalog                     |  26987 | 392.5 | 394.3 |    0% | 361.8 | 362.4 |    0% |
| random                           |   7976 | 708.8 | 714.6 |    0% | 636.1 | 638.1 |    0% |

## Things I Didn't Do

For the record, here are the things I wanted to do but couldn't find a way without hurting simdjson core performance:

* **Remove `increment_count()`.** Users should just increment when they receive a primitive() or start_array() call. But moving increment_count() in *any way* kills performance. In particular, moving [this one](https://github.com/simdjson/simdjson/pull/1101/files#diff-73e01dfaa9e28fe806176f82736d91ccR124) literally anywhere else causes a 1-9% performance drop across our test files, and that's the one that we'd really need to move because it's wedged in between several independent tasks. Given that it's basically a completely independent operation, this is super puzzling.

  ![image](https://user-images.githubusercontent.com/654882/89832401-1d8f9780-db14-11ea-8e5e-f623de32451b.png)

  More puzzling, that one can be removed entirely by setting container count to 1 when you first create the open_container in start_object(). Removing it has the same performance effect as moving it (even just removing it and changing nothing else has this effect).
* **Pass key and value together (i.e. `primitive_field(key, value)`).** This is related to the first one, actually. I'd like to keep the key around and pass it along with the value so the user can pump them both into a hashtable in one go. This is related to the first one, actually: if we could move that increment_count around, I think we could safely move key parsing straight into the `object_field` goto. That would also save us some code (at the least, there'd be one key string parsing routine instead of two!).
* **Remove STREAMING template parameter from walk_document.** There are only two statements that respect streaming, one at the beginning and one at the end. Somehow the first one, of the form `if (!STREAMING) { if (last structural == ']') { ... } }`, causes a performance drop even if you just remove `if (!STREAMING)` (a template parameter!). The if (!STREAMING) at the very end can't be moved out of the method without a performance drop, either.
* **Let user consume arrays/objects themselves.** I experimented with a `visit_value()` that is called for any value (including `{` and `[`): you would return either state::handled, state::array or state::object, and the core algorithm would branch from there. This gives you the freedom to skip or even parse specific objects yourself, allowing the tweet reader (for example) to skip all objects and arrays that don't have the stuff it needs, as well as run faster by expecting fields in a specific order instead of hashing. Those benefits are theoretical though, and depend on on-demand parsing working (it's an interplay between them), so I don't feel terrible about waiting on this one. On ARM I got a version that ran a little faster on some files, actually, but on Intel I couldn't keep it from going at least a little negative. I could have been screwed by jitters, but I don't want to chance that just now.