@lemire 2020-08-07
On my todo