@jkeiser 2020-08-07
For the record, here are the things I wanted to do but couldn't:

* **Remove `increment_count()`.** Users should just increment when they receive a primitive() or start_array() call. But moving increment_count() in *any way* kills performance. In particular, moving [this one](https://github.com/simdjson/simdjson/pull/1101/files#diff-73e01dfaa9e28fe806176f82736d91ccR124) literally anywhere else causes a 1-9% performance drop across our test files, and that's the one that we'd really need to move because it's wedged in between several independent tasks. Given that it's basically a completely independent operation, this is super puzzling.
* **Pass key and value together (i.e. `primitive_field(key, value)`).** This is related to the first one, actually. I'd like to keep the key around and pass it along with the value so the user can pump them both into a hashtable in one go. This is related to the first one, actually: if we could move that increment_count around, I think we could safely move key parsing straight into the `object_field` goto. That would also save us some code (at the least, there'd be one key string parsing routine instead of two!).
* **Remove STREAMING template parameter from walk_document.** There are only two statements that respect streaming, one at the beginning and one at the end. Somehow the first one, of the form `if (!STREAMING) { if (last structural == ']') { ... } }`, causes a performance drop even if you just remove `if (!STREAMING)` (a template parameter!). The if (!STREAMING) at the very end can't be moved out of the method without a performance drop, either.