@lemire 2020-08-11
I'll try to write a little experiment. It is hard to assess an API such as this one without trying it out a bit.