@jkeiser 2020-08-11
To be clear, I don't consider this a final SAX API. This is "as close as I can get to a public SAX API while preserving stage 2 performance."

There are still hard tradeoffs to be made (probably) before reaching a satisfying API.

That said, it's not *bad* as far as it goes and I encourage you to try it out. The SAX benchmark test does use this API, if you want an example of non stage 2 usage.