@lemire 2020-08-11
Oh. I understand. 

It is precisely because it is such interesting work that I want to do more than just read it.