@lemire 2020-08-12
What is clear from the examples, starting with your own, is that we need "typed" number parsing. We don't need it in the DOM cases, but it is obvious that SAX or on-demand is going to require typed numbers.


Then there is the little issue of testing these new functions.

We might as well do it properly.


Regarding making this API public, I think we should definitively hold on. It is good for our internal tests and tuning, but it is clearly an inconvenient API.