@jkeiser 2020-08-13
I have updated the API to be a little better, with the same caveats as before--specifically, this is all the changes I can make without affecting performance negatively. In fact, the current PR seems to help a little bit on the performance front.

There's more I want, and we'll definitely address that stuff in a separate PR. Skip functionality would be awesome and help clean up the SAX consumer a lot.