@jkeiser 2020-08-13
I just tried removing the (unnecessary) number parsing changes I made, which sets performance at something of an even keel. The number parsing changes, turns out, help with this patch, but make master worse. Such is stage 2 life.