@jkeiser 2020-08-13
I'll make a new PR with the number changes, even though they improve performance when added to this one. They are unrelated and this is big enough as it is. I'd like this checkin to stick to the rule "not changing logic in any way whatsoever."