@jkeiser 2020-08-18
@lemire I believe this one is ready for review, at your leisure.