@lemire 2020-08-18
Yes. It is on my todo.