@lemire 2020-08-18
I have reverified that this is not a negative performance-wise.