@jkeiser 2020-08-19
I just rebased everything against master. I have pushed the pre-rebase code to the [yakety-sax-pre-rebase](https://github.com/simdjson/simdjson/tree/jkeiser/yakety-sax-pre-rebase) branch, just in case we notice something off.