@jkeiser 2020-08-19
Final numbers vs. master are frustratingly hard to match up on Intel--not just because master moved, but also because when this PR is put on top of the new master it increases instruction count in more files. Note that these numbers also differ because they include the number patches from #1114.

ARM, however, is in roughly the same spot as before.

### Haswell gcc10.2 skylake

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| canada                           |  35172 | 292.2 | 281.9 |    3% | 980.7 | 958.6 |    2% |  18736 |  18990 |   -1% |  40598 | 116137 | -186% | 344079 | 331265 |    3% |
| mesh.pretty                      |  24646 | 183.8 | 182.5 |    0% | 593.1 | 586.5 |    1% |  13256 |  13458 |   -1% |      0 |    425 |    0% | 194601 | 181070 |    6% |
| gsoc-2018                        |  51997 |  71.5 |  71.1 |    0% | 167.1 | 168.0 |    0% |  29835 |  30923 |   -3% |  94882 |  89192 |    5% | 363303 | 350433 |    3% |
| apache_builds                    |   1988 |  85.8 |  86.1 |    0% | 303.4 | 306.1 |    0% |    104 |    102 |    1% |      0 |      0 |    0% |  14139 |  13316 |    5% |
| semanticscholar-corpus           | 134271 | 108.4 | 109.4 |    0% | 283.4 | 285.7 |    0% |  98053 |  98756 |    0% | 621251 | 634300 |   -2% | 1207637 | 1188039 |    1% |
| twitterescaped                   |   8787 | 185.2 | 187.3 |   -1% | 488.6 | 492.4 |    0% |   5143 |   4922 |    4% |      0 |      0 |    0% |  79723 |  74377 |    6% |
| instruments                      |   3442 | 105.9 | 107.1 |   -1% | 378.8 | 382.1 |    0% |    463 |    475 |   -2% |      4 |      0 |    0% |  30035 |  29251 |    2% |
| github_events                    |   1017 |  78.9 |  79.8 |   -1% | 263.5 | 266.1 |   -1% |    117 |    123 |   -5% |      0 |      0 |    0% |   2727 |   2370 |   13% |
| random                           |   7976 | 140.0 | 141.7 |   -1% | 492.3 | 497.7 |   -1% |   1999 |   1919 |    4% |      0 |      0 |    0% |  91848 |  86378 |    5% |
| marine_ik                        |  46616 | 299.2 | 303.2 |   -1% | 918.1 | 919.1 |    0% |  53108 |  53593 |    0% | 279405 | 279437 |    0% | 569847 | 557580 |    2% |
| update-center                    |   8330 | 110.5 | 112.2 |   -1% | 332.9 | 337.5 |   -1% |   5529 |   5521 |    0% |      0 |      0 |    0% |  84293 |  79052 |    6% |
| mesh                             |  11306 | 312.6 | 318.7 |   -1% | 1011.1 | 1007.1 |    0% |  10051 |  10035 |    0% |      0 |      0 |    0% | 137021 | 127307 |    7% |
| numbers                          |   2345 | 264.3 | 269.7 |   -2% | 811.3 | 824.1 |   -1% |   2112 |   2146 |   -1% |      0 |      2 |    0% |  18512 |  17565 |    5% |
| twitter                          |   9867 |  90.2 |  92.2 |   -2% | 288.8 | 292.1 |   -1% |   3476 |   3521 |   -1% |      0 |      0 |    0% |  83426 |  78145 |    6% |
| citm_catalog                     |  26987 |  80.4 |  82.6 |   -2% | 296.8 | 297.5 |    0% |   1731 |   1741 |    0% |     17 |      0 |    0% | 194983 | 181754 |    6% |

### ARM64 gcc10.2 ampere

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| canada                           |  35172 | 813.0 | 776.2 |    4% | 994.2 | 985.6 |    0% |  48885 |  47923 |    1% |  91694 |  91695 |    0% | 5088709 | 4982458 |    2% |
| mesh.pretty                      |  24646 | 562.7 | 553.3 |    1% | 624.4 | 622.3 |    0% |  24134 |  23190 |    3% |  59097 |  59086 |    0% | 1975343 | 1932569 |    2% |
| citm_catalog                     |  26987 | 393.3 | 389.5 |    0% | 362.8 | 362.4 |    0% |   8689 |  10139 |  -16% |  62686 |  62671 |    0% | 1812840 | 1816819 |    0% |
| semanticscholar-corpus           | 134271 | 493.0 | 489.1 |    0% | 381.2 | 381.1 |    0% | 203552 | 197893 |    2% | 314160 | 314340 |    0% | 12953899 | 12963555 |    0% |
| apache_builds                    |   1988 | 471.2 | 468.0 |    0% | 412.5 | 412.1 |    0% |    694 |    691 |    0% |   4830 |   4826 |    0% | 201028 | 200962 |    0% |
| update-center                    |   8330 | 541.7 | 539.6 |    0% | 456.3 | 456.1 |    0% |  13636 |  13163 |    3% |  20801 |  20793 |    0% | 1020203 | 1020477 |    0% |
| instruments                      |   3442 | 492.0 | 490.3 |    0% | 465.5 | 465.2 |    0% |   4558 |   4410 |    3% |   8677 |   8666 |    0% | 341150 | 340877 |    0% |
| twitter                          |   9867 | 488.4 | 487.1 |    0% | 382.9 | 382.8 |    0% |  16530 |  16263 |    1% |  23414 |  23408 |    0% | 856508 | 854519 |    0% |
| random                           |   7976 | 700.6 | 699.7 |    0% | 637.1 | 636.6 |    0% |   6824 |   7811 |  -14% |  21582 |  21586 |    0% | 1235816 | 1238507 |    0% |
| twitterescaped                   |   8787 | 734.9 | 734.6 |    0% | 621.1 | 621.0 |    0% |  19266 |  19116 |    0% |  21341 |  21368 |    0% | 1560409 | 1557890 |    0% |
| gsoc-2018                        |  51997 | 359.5 | 359.5 |    0% | 238.6 | 238.5 |    0% |  39149 |  39449 |    0% | 109739 | 109645 |    0% | 2750274 | 2752790 |    0% |
| github_events                    |   1017 | 459.0 | 459.2 |    0% | 358.0 | 357.8 |    0% |   1512 |   1526 |    0% |   2405 |   2403 |    0% |  87555 |  87581 |    0% |
| mesh                             |  11306 | 916.9 | 917.7 |    0% | 1045.8 | 1047.2 |    0% |  20724 |  18642 |   10% |  32333 |  32339 |    0% | 1649235 | 1611253 |    2% |
| marine_ik                        |  46616 | 856.6 | 859.8 |    0% | 967.0 | 967.7 |    0% |  95245 |  87452 |    8% | 133984 | 133930 |    0% | 7549658 | 7441098 |    1% |
| numbers                          |   2345 | 889.2 | 899.0 |   -1% | 870.7 | 874.9 |    0% |   3270 |   3249 |    0% |   6018 |   6013 |    0% | 258837 | 247087 |    4% |